requirements.txt
python-telegram-bot==21.6
anthropic
python-dotenv
